check delete

+++ src/main/java/org/example/uni_style_be/service/impl/MaterialServiceImpl.java
@@ -6,12 +6,14 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.example.uni_style_be.entities.Material;
+import org.example.uni_style_be.enums.InvalidInputError;
 import org.example.uni_style_be.enums.NotFoundError;
 import org.example.uni_style_be.exception.ResponseException;
 import org.example.uni_style_be.model.filter.MaterialParam;
 import org.example.uni_style_be.model.request.MaterialRequest;
 import org.example.uni_style_be.model.response.MaterialResponse;
 import org.example.uni_style_be.repositories.MaterialRepository;
+import org.example.uni_style_be.repositories.ProductDetailRepository;
 import org.example.uni_style_be.repositories.specification.MaterialSpecification;
 import org.example.uni_style_be.service.MaterialService;
 import org.springframework.data.domain.Page;
@@ -29,6 +31,7 @@ public class MaterialServiceImpl implements MaterialService {
 
     MaterialRepository materialRepository;
     ObjectMapper objectMapper;
+    ProductDetailRepository productDetailRepository;
 
     @Override
     @Transactional
@@ -56,6 +59,9 @@ public class MaterialServiceImpl implements MaterialService {
 
     @Override
     public void delete(Long id) {
+        if (productDetailRepository.existsByMaterial_Id(id)) {
+            throw new ResponseException(InvalidInputError.MATERIAL_HAS_PRODUCT);
+        }
         materialRepository.deleteById(id);
     }
